+++ src/Obj/ObjType.php
@@ -7,6 +7,7 @@ namespace Serhii\GoodbyeHtml\Obj;
 enum ObjType: string
 {
     case INTEGER_OBJ = 'INTEGER';
+    case FLOAT_OBJ = 'FLOAT';
     case ERROR_OBJ = 'ERROR';
     case STRING_OBJ = 'STRING';
     case HTML_OBJ = 'HTML';