+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Changed the first parameter of the `Parser` class. Now, it accepts absolute path to the template file or a string with template content. Before, it was accepting only the path
 - Improved code readability to refactoring the `Parser` class
+- Added support for `floats` to the lexer and core parser
 
 ----